requirements.txt
flask
Flask-Session
gunicorn
jinja2
pandas
requests
python-dotenv
tator
